requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-Mail==0.9.1
Werkzeug==2.3.7
MySQLdb==1.4.6
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
waitress==2.1.2

# Cart Abandonment Detection System
openai>=1.0.0
scikit-learn==1.3.2
numpy==1.26.2